Day 11 part 2 optimized
This greatly reduces the number of array iterations and gets rid of
all `in_array()` checks, making the algorithm much faster than before.
On first my first take, getting answers to both parts of the puzzle
required around 1.5 seconds, of which around 0.9 seconds was used for
actual code executio. Now the answers are returned in around 0.6 sec,
of which around only 0.015 secs is required for actual code execution.