Dec 2 2024
This took me a lot more than it should have. I had a bug which didn't pop up
with the test data, nor with my modified test data. I ended up trying with the
same (larger) `try_remove_indexes` definition for both checks, which,
apparently, worked. When that worked, I figured out by comparing debug output
that the problem was me not attempting removal of level #1 (0-based) when
difference between level #1 and level #2 was bad, which meant that I missed
some of the cases where level #1 was out of the "dynamic" of the report
(e.g. [7, 4, 8, 10, 12]).