Dec 6 2024, failed optimization
Not only did the algo turn out slower, it also produces incorrect result for part 2.
After spending a big part of the day on it, and having a look at someone else's Python
implementation which took like 4 times fewer lines and does its job in 4 seconds
(that's over 100 times faster), I just gave up on the idea of improving *this* code
further. It should be thrown away and rewritten, but realistically, I doubt I'll ever
do that. So, committing this as-is just to remember having tried.